Set Float to type f64

+++ src/types.rs
@@ -3,5 +3,5 @@
 
 // PAWS types
 
-pub type Float = i64;
+pub type Float = f64;
 pub type Int = i32;
